Re-set osc.spec.criconfig on re-deploy
Now that it is possible to modify the CRI for existing worker pools, the osc needs to take into account that a worker pool formerly set to `containerd` is reconfigured with `worker.CRI==nil`. The OSC object is only re-created when k8s major or minor version is changed or the worker name is updated and therefore needs to take this update path into account.